requirements.txt
airflow
qrcode
pillow
requests